fix(admin): harden directory command support (#403) (thanks @dl-alexandre)
